+++ spec/dummy_app/app/models/kitchen/banana.rb
@@ -2,6 +2,6 @@
 
 module Kitchen
   class Banana < ActiveRecord::Base
-    has_paper_trail class_name: "Kitchen::BananaVersion"
+    has_paper_trail versions: { class_name: "Kitchen::BananaVersion" }
   end
 end