+++ spec/dummy_app/app/models/post.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class Post < ActiveRecord::Base
-  has_paper_trail class_name: "PostVersion"
+  has_paper_trail versions: { class_name: "PostVersion" }
 end