+++ CHANGELOG.md
@@ -18,6 +18,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
+- [#1166](https://github.com/paper-trail-gem/paper_trail/pull/1166) -
+  New global option `has_paper_trail_defaults`, defaults for `has_paper_trail`
 - [#1158](https://github.com/paper-trail-gem/paper_trail/pull/1158) — Add the
   ability to pass options, such as `scope` or `extend:` to the `has_many
   :versions` association macro.