+++ CHANGELOG.md
@@ -23,6 +23,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 - [#1158](https://github.com/paper-trail-gem/paper_trail/pull/1158) — Add the
   ability to pass options, such as `scope` or `extend:` to the `has_many
   :versions` association macro.
+- [#1171](https://github.com/paper-trail-gem/paper_trail/pull/1171) -
+  Remove the upperbound constraint on ActiveRecord.
 
 ### Fixed