+++ CHANGELOG.md
@@ -36,7 +36,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- [#1159](https://github.com/paper-trail-gem/paper_trail/pull/1159) -
+  Support rails 6.0.0-alpha
 
 ### Fixed